+++ src/main/java/com/pooli/family/domain/dto/response/SharedDataThresholdResDto.java
@@ -12,6 +12,6 @@ public class SharedDataThresholdResDto {
     @Schema(description = "가족 공유 데이터 임계치 설정 여부", example = "true")
     private Boolean isThresholdActive;
     @Schema(description = "가족 공유 데이터 사용량 임계치 값(단위: Byte)", example = "1000")
-    private Integer familyThreshold;
+    private Long familyThreshold;
 
 }
